+++ src/dsl/timeIntegration/forwardEuler.cpp
@@ -3,9 +3,8 @@
 
 #include <memory>
 
-#include "NeoFOAM/finiteVolume/cellCentred/timeIntegration/forwardEuler.hpp"
+#include "NeoFOAM/dsl/timeIntegration/forwardEuler.hpp"
 #include "NeoFOAM/core/error.hpp"
-#include "NeoFOAM/core/parallelAlgorithms.hpp"
 
 namespace NeoFOAM::finiteVolume::cellCentred
 {
