+++ test/finiteVolume/CMakeLists.txt
@@ -4,4 +4,3 @@
 add_subdirectory(cellCentred/fields)
 add_subdirectory(cellCentred/boundary)
 add_subdirectory(cellCentred/interpolation)
-add_subdirectory(cellCentred/timeIntegration)
